requirements.txt
arabic-reshaper==2.1.4
asn1crypto==1.5.1
autocommand==2.2.1
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cheroot==8.6.0
CherryPy==18.8.0
click==8.1.3
cryptography==38.0.1
cssselect2==0.7.0
decorator==5.1.1
docopt==0.6.2
et-xmlfile==1.1.0
fastrlock==0.8
future==0.18.2
grpcio==1.49.1
html5lib==1.1
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
inflect==6.0.0
ipython==7.34.0
jaraco.classes==3.2.3
jaraco.collections==3.5.2
jaraco.context==4.1.2
jaraco.functools==3.5.2
jaraco.text==3.9.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
kaleido==0.2.1
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
minknow-api==5.0.0.1
more-itertools==8.14.0
numpy==1.21.6
openpyxl==3.0.10
oscrypto==1.3.0
packaging==21.3
pandas==1.3.5
parso==0.8.3
patsy==0.5.3
pdf2image==1.16.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pipreqs==0.4.11
plotly==5.10.0
portend==3.1.0
prompt-toolkit==3.0.31
protobuf==3.20.3
psutil==5.9.2
ptyprocess==0.7.0
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
pyHanko==0.15.0
pyhanko-certvalidator==0.19.5
pyparsing==3.0.9
PyPDF3==1.0.6
pyRFC3339==1.1
pysam==0.19.1
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.11
requests==2.28.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
statsmodels==0.13.2
svglib==1.4.1
tempora==5.0.2
tenacity==8.1.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tqdm==4.64.1
traitlets==5.4.0
typing_extensions==4.4.0
tzdata==2022.5
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
xhtml2pdf==0.2.8
yarg==0.1.9
zc.lockfile==2.0
zipp==3.9.0
